fix(vitrine): update the wasm JSX test call sites for the compat argument
