+++ crates/vize_carton/src/config/loader/tests.rs
@@ -166,83 +166,6 @@ fn load_compiler_vue_version_defaults_to_unset_for_vue3() {
 }
 
 #[test]
-fn load_compiler_jsx_mode_reads_jsx_mode_key() {
-    let dir = tempfile::tempdir().unwrap();
-    let config_path = dir.path().join("vize.config.json");
-    std::fs::write(&config_path, r#"{ "compiler": { "jsxMode": "vapor" } }"#).unwrap();
-
-    assert_eq!(
-        load_compiler_jsx_mode(Some(&config_path)),
-        Some(JsxMode::Vapor)
-    );
-    assert_eq!(
-        load_compiler_jsx_mode(Some(&config_path)).map(JsxMode::as_str),
-        Some("vapor")
-    );
-}
-
-#[test]
-fn load_compiler_jsx_mode_reads_vdom() {
-    let dir = tempfile::tempdir().unwrap();
-    let config_path = dir.path().join("vize.config.json");
-    std::fs::write(&config_path, r#"{ "compiler": { "jsxMode": "vdom" } }"#).unwrap();
-
-    assert_eq!(
-        load_compiler_jsx_mode(Some(&config_path)),
-        Some(JsxMode::Vdom)
-    );
-}
-
-#[test]
-fn load_compiler_jsx_mode_defaults_to_unset() {
-    let dir = tempfile::tempdir().unwrap();
-    let config_path = dir.path().join("vize.config.json");
-    std::fs::write(&config_path, r#"{ "formatter": { "singleQuote": true } }"#).unwrap();
-
-    // No `compiler.jsxMode` key → absent (the JSX entry points treat this as VDOM).
-    assert_eq!(load_compiler_jsx_mode(Some(&config_path)), None);
-}
-
-#[test]
-fn load_compiler_jsx_compat_reads_babel_key() {
-    let dir = tempfile::tempdir().unwrap();
-    let config_path = dir.path().join("vize.config.json");
-    std::fs::write(&config_path, r#"{ "compiler": { "jsxCompat": "babel" } }"#).unwrap();
-
-    assert_eq!(
-        load_compiler_jsx_compat(Some(&config_path)),
-        Some(JsxCompat::Babel)
-    );
-    assert_eq!(
-        load_compiler_jsx_compat(Some(&config_path)).map(JsxCompat::as_str),
-        Some("babel")
-    );
-}
-
-#[test]
-fn load_compiler_jsx_compat_reads_native() {
-    let dir = tempfile::tempdir().unwrap();
-    let config_path = dir.path().join("vize.config.json");
-    std::fs::write(&config_path, r#"{ "compiler": { "jsxCompat": "native" } }"#).unwrap();
-
-    assert_eq!(
-        load_compiler_jsx_compat(Some(&config_path)),
-        Some(JsxCompat::Native)
-    );
-}
-
-#[test]
-fn load_compiler_jsx_compat_defaults_to_unset() {
-    let dir = tempfile::tempdir().unwrap();
-    let config_path = dir.path().join("vize.config.json");
-    std::fs::write(&config_path, r#"{ "compiler": { "jsxMode": "vdom" } }"#).unwrap();
-
-    // No `compiler.jsxCompat` key → absent, which the JSX entry points treat as
-    // `native`. Setting only `jsxMode` must not imply a compatibility mode.
-    assert_eq!(load_compiler_jsx_compat(Some(&config_path)), None);
-}
-
-#[test]
 fn load_config_uses_typescript_config() {
     let dir = tempfile::tempdir().unwrap();
     let config_path = dir.path().join("vize.config.ts");
